Split up assertions into multiple test classes
