+++ convert-junit4-to-junit5/src/test/java/jb/convert/ast/AssertMigrationTest.java
@@ -1,0 +1,98 @@
+package jb.convert.ast;
+
+import org.junit.jupiter.api.Test;
+
+import static jb.JunitConversionLogicFixture.assertAfterAddingClassAfter;
+import static jb.JunitConversionLogicFixture.assertAfterWrappingInMethod;
+import static jb.JunitConversionLogicFixture.assertUnchangedAfterWrappingInMethod;
+import static jb.JunitConversionLogicFixture.assertWrappingInClass;
+
+class AssertMigrationTest {
+
+    @Test
+    void importWildcardStaticAssertions() {
+        String code = "import static org.junit.Assert.*;";
+        String expected = "import static org.junit.jupiter.api.Assertions.*;";
+        assertAfterAddingClassAfter(code, expected);
+    }
+
+    @Test
+    void specificStaticAssertions() {
+        String originalImport = "import org.junit.Assert;";
+        String originalMethod = "Assert.assertTrue(a,b);";
+        String expectedImport = "import org.junit.jupiter.api.Assertions;";
+        String expectedMethod = "Assertions.assertTrue(b,a);";
+        assertAfterWrappingInMethod(originalImport, originalMethod, expectedImport, expectedMethod);
+    }
+
+    @Test
+    void specificStaticFail() {
+        String originalImport = "import org.junit.Assert;";
+        String originalMethod = "Assert.fail(a);";
+        String expectedImport = "import org.junit.jupiter.api.Assertions;";
+        String expectedMethod = "Assertions.fail(a);";
+        assertAfterWrappingInMethod(originalImport, originalMethod, expectedImport, expectedMethod);
+    }
+
+    @Test
+    void specificFullyQualifiedStaticAssertions() {
+        String originalImport = "import org.junit.Assert;";
+        String originalMethod = "org.junit.Assert.assertTrue(a,b);";
+        String expectedImport = "import org.junit.jupiter.api.Assertions;";
+        String expectedMethod = "org.junit.jupiter.api.Assertions.assertTrue(b,a);";
+        assertAfterWrappingInMethod(originalImport, originalMethod, expectedImport, expectedMethod);
+    }
+
+    @Test
+    void importWildcardMainPackage() {
+        String originalImport = "import org.junit.*;";
+        String expectedImport = "import org.junit.jupiter.api.*;";
+        assertAfterAddingClassAfter(originalImport, expectedImport);
+    }
+
+    @Test
+    void specificTestAnnotation() {
+        String annotation = "import org.junit.Test;";
+        String expected = "import org.junit.jupiter.api.Test;";
+        assertAfterAddingClassAfter(annotation, expected);
+    }
+
+    @Test
+    void singleAssertion() {
+        String code = "assertTrue(message, actual);";
+        String expected = "assertTrue(actual, message);";
+        assertAfterWrappingInMethod(code, expected);
+    }
+
+    @Test
+    void multipleAssertions() {
+        String code = "assertTrue(message, actual);";
+        String expected = "assertTrue(actual, message);";
+        assertAfterWrappingInMethod(code, expected);
+    }
+
+    @Test
+    void nonJunitMethodCall() {
+        String code = "doWork(message, actual);";
+        assertUnchangedAfterWrappingInMethod(code);
+    }
+
+    // -------------------------------------------------------
+
+    @Test
+    void assertInFirstLineOfMethod() {
+        String code = "public void method() { assertTrue(message, actual); }";
+        String expected = "public void method() { assertTrue(actual, message); }";
+        assertWrappingInClass(code, expected);
+    }
+
+    @Test
+    void assertAfterCommentOnPreviousLine() {
+        String code = "public void method() { assertTrue(message, actual); // comment \n"
+                + "assertTrue(message, actual); }";
+        String expected = "public void method() { assertTrue(actual, message); // comment \n"
+                + "assertTrue(actual, message); }";
+        assertWrappingInClass(code, expected);
+    }
+
+}
